Separate the Panels into their own functions. Config now controls writing itself. App is now solely concerned with app level logic
