+++ src/ui/mod.rs
@@ -2,7 +2,12 @@
 //! and interface state management
 pub mod app;
 mod config;
+mod config_panel;
+mod ecc_panel;
+mod error;
 mod graph_manager;
+mod graph_panel;
 mod rate_graph;
+mod router_panel;
 mod status_colors;
 mod time_format;
